#48 refactor: geolocation 함수를 메인페이지와 검색페이지에서 각각 불러와야 하기 때문에 customHook으로 관리
